requirements.txt
python-dotenv
langchain
langchain-core
langchain-openai
langgraph
tavily-python
amadeus
weatherapi
streamlit
